VPNaaS Service Driver for Cisco CSR
This has the service driver part of the vendor specific VPNaaS plugin.
The changes rely on the Service Type Framework code, which is presently
under review (client 53602, server 41827). The device driver will be
under a separate review.

Note: The Cisco plugin will require an out-of-band CSR setup and
      running for operation.

Note: See review 74156 for more details on device driver portion of
      this blueprint.

Change-Id: I39b1475c992b594256f5a28be0caa1ee9398050e
Partially-implements: blueprint vpnaas-cisco-driver