Makes the Extension loader behavior predictable
When using a cluster of load-balanced Neutron Servers,
the order of extension files being loaded may differ
from one server to another. This is usually harmless,
but it is better to force the list to be loaded in the
same way across the entire cluster, just in case
something funky is going on.

Partial-bug: #1285999

Change-Id: Ib0bfd24fc52b6cd9f90c350d3af496f04a90f50a