Mock resource attribute dicts in tests
Previously, setup for neutron unit tests might include saving a copy
of resource attribute dicts so that they could be restored to a
pristine state on teardown.  This patch adds a helper method on the
base test class that uses mock.patch.dict() to accomplish the same
goal more cleanly, and ensures that test classes only mock if their
parent classes have not yet done so.

Change-Id: I4002e96b5aea9375346ac1d13e756e5ee5494ead