Improved quota error message
Removed two negatives and showed a clear message when
tried to see quota of different tenants.
When a user who does not belong to a project tries to access the
quota of that project, the message that gets displayed is not clear.
In this change the user would understand the error more properly.

Change-Id: I2284df07687bb530fe06fbaab38e2971826b7b40
Closes-Bug: #1288915