Typographical correction of Arista ML2 help
The Configuration Reference Guide on docs.openstack.org contains
configuration tables generated based on the help strings defined
in the Python code.

Typographical errors in the source for the Arista ML2 plugin,
particularly a lack of spaces between the start and end of
continuing lines, were resulting in errors in the guide.

Change-Id: I793274d9bf631db890af57f242b99de99db6f1f6