Fix missing reference service_provider for VPNaaS
In Neutron's stable/havana branch, the sample neutron.conf file
is missing a service_providers line for the reference
open source backend for VPNaaS (OpenSwan).  Since packaging relies
on this to provide a default value, VPNaaS doesn't work out of the
box in Havana.  This has been corrected upstream in Icehouse, but
the fix looks unlikely to be backported to stable/havana at this
point so we'll carry it locally.

Partial-Bug: #1301418