Verify subnet configuration has valid CIDR & GW
This patch is a forward porting of three closely related patches
from the puppet-coe-service-patch module that are intended
to block creation of /30, /31, and /32 subnets.  The commit log
messages from the original patches are below.  These are partly
related to upstream bug #1341040 for which the community has
adopted a different and less strict approach.  We have opted
for a stricter approach, so we'll need to carry this patch for now.

---
commit 8dabd46a2f9827e9b10dfe3461a986392ad7d82f
Author: Pauline Yeung <yeungp@cisco.com>
Date:   Thu Jul 24 15:45:07 2014 -0700

    DE631 block /30 in neutron server, in addition to blocking in
    CLI and dhcp an addition to blocking in CLI and dhcp agent.

    Change-Id: If405e49b8faa4c6a584441ee93dab211a8f521e6
---
---
commit dff604420c48dfe47467cdd1b48687443fb7b32a
Author: Pauline Yeung <yeungp@cisco.com>
Date:   Wed Jul 23 17:49:45 2014 -0700

    DE631 do not allow /30, as creating the first VM may require network
    with at least 5 IP addresses.

    Change-Id: I1bb48a512adeb509073b8b1c61520db8ea555768
---
---
commit 3a3d4e95720846a2ed3439c46bf33407da08a2d2
Author: Pauline Yeung <yeungp@cisco.com>
Date:   Thu Jul 17 17:16:20 2014 -0700

    DE494 add CIDR and gateway IP to payload, data which always exists
    in a subnload, data which always exists in a subnet.

    Change-Id: I30121506fb918f6e5932e627d07067d289194a9d

---

Partial-bug: #1341040
Partial-rally-bug: rally-bug DE631
Partial-rally-bug: rally-bug DE494
Implements: rally-user-story US4145
Implements: rally-task TA3473
Not-in-upstream: true