Use information from the dnsmasq hosts file to call dhcp_release
Certain situations can cause the DHCP agent's local cache to get out
of sync with the leases held internally by dnsmasq.  This method of
detecting when to call dhcp_release is idempotent and not dependent on
the cache.  It is more robust.

Closes-Bug: #1250644
Implements: rally-user-story US4052
Implements: rally-task TA4018
Upstream-Review: https://review.openstack.org/#/c/56263
Not-in-upstream: false