travis-ci: Change to run_tests.sh and chance concurrency
Tests on travis-ci are being killed due to OOM. The number of worker
threads for testr are large and this change make it one so that memory
requirement is smaller and tests run serially and pass. The second
problem is that even adding travis_wait doesn't allow tests to complete
in 20 minutes as there is not output for tox, so use run_test.sh
instead.

Change-Id: I9be1587c01ee3fbd049faf84f493fe8b7da6bc29