Apply 'int allowed vlan' cfg when one not there
This fixes a bug introduced by 1330597 which expects all Nexus devices
to return an error message when 'switchport trunk allowed vlan add' is
applied and this CLI currently is not applied.  Instead, code was
changed to get current interface config and check if this config is
not there.  It config currently applied, then 'add' keyword is used
in next config; otherwise, it is omitted.

Change-Id: Id62d3fe96f175ce619540c681ec164022cc03f0a
Closes-bug:  #1395122