Backport https://review.openstack.org/#/c/108272/
This should adddress the DHCPNAK issue we are seeing
in Havan (and likely Icehouse as well).

This code will cause dnsmasq to execute dhcp-script
which will readin the hosts files and create leases
for everything in them, thus elininating the DHCPNAK
which comes from the dnsmasq process no longer knowing
about the lease.

Closes-rally-bug: DE1821
Upstream-Review: https://review.openstack.org/#/c/108272/5
Not-in-upstream: true