Support Multiple IPv6 Prefixes and Addresses for IPv6 Network
This change adds functional code and UT for the following work items
as listed in the Neutron blueprint 'multiple-ipv6-prefixes'.
- Port-create API handling changes
- Port-update API handling changes

This does not include changes to the number of IPv6 prefixes and
addresses supported on gateway interfaces and internal router
interfaces. This support will be added with two followup commit.

Change-Id: I182b8874b40fcc6ff6662b9a785ec4f487a732cb
Partially-implements: blueprint multiple-ipv6-prefixes