Include IPv6 SLAAC addresses implicitly for port create/update
(Patch set #1 for the multiple-ipv6-prefixes blueprint)

This change adds functional code and UT for the following work items
as listed in the Neutron blueprint 'multiple-ipv6-prefixes'.
- Port-create API handling changes
- Port-update API handling changes

This does not include changes to the number of IPv6 prefixes and
addresses supported on gateway interfaces and internal router
interfaces. This support will be added with two followup commits.

Change-Id: I61beb2d663dfa6070b529b567d79664b1cb69810
Partially-implements: blueprint multiple-ipv6-prefixes