CSCut07075 Send tenant_id 0 when network updated to be shared
For L3 forwarding we need to indicate if a network is shared. In order
to do so, we will now send tenant_id = '0' for shared networks. This
change additionally ensures the tenant_id is properly updated when the
network is updated to be a shared network after it is created.

Change-Id: I50c0d19a47f82994a9b17fc70ea9e2fba6deec39
Closes-CDETS-Bug: CSCut07075