Wrong profile in port create causes stale db entries
During port-creates, if you provide an incorrect name or UUID for
the policy profile, it results in a Mechanism Driver Exception but
still continues with the port create causing a PortBinding
Exception. This patch adds a check in extension driver causing the
port create to fail and not lead to stale entries in the db.

Closes-bug: CSCus91614