CSCut18124 Brcd_3: Implement update port
Added support for update port operation in order to have this operation
work successfully. On update port we check if this the the first time
binding of a router or DHCP port as currently that is the only operation
we want to support updates for.

Change-Id: I69624d7b0cb0eb82bc5cf6ed19202e076bee7cc4
(cherry picked from commit fd95a47756478f5858cf5c176b2dff9f3fe679ad)