Avoid changes to core neutron files
Previously a new RPC topic and some service plugin
constants were defined in files part of neutron core.

This patch moves those definitions to cisco files
which removes the need to upstream those definitions
to the community neutron repo. Instead the definitions
will follow with the rest of the cisco code to the
networking-cisco repo.

This patch also fixes a few pep8 failures.

Change-Id: Iae92d754f9cd697a56591a8119b2fb17ba9e5103