Add plugging driver for VLAN trunking using N1kv ML2 mechanism driver.
As of Liberty the N1kv monolithic plugin has been deleted. Instead
the N1kv mechanism driver for ML2 is to be used. This patch adds
a plugging driver that works with that N1kv ML2 mechanism driver.
That plugging driver uses the N1kv VLAN trunking capability to
provide connectivity for Neutron routers hosted in CSR1kv VMs.

The patch also provides unit tests for the new plugging driver.

Change-Id: If00abb74c2817d3f360266f8cfc76a59094c6d78