Allow non-neutron management network for hardware hosting devices
So far the management network for hosting device must have been
a neutron network. This patch removes that requirement.

Note that for a deployment with virtual machine hosting devices
or with a mix of hardware and virtual machine hosting devices,
the management network for hosting devices must still be a
neutron network.

Change-Id: Idafb189e02395b7f3a00ae56c72dd1a51be205e4