Add type and service for logical resource slot allocations
The uuid of a logical resource (e.g., a Neutron router) is
not stored as in the slot allocations DB table as a foreign
key to the logical resource's table. The reasons is that
different type of logical resources must be possible to store.

To make it easier to determine which type of resource a
particular uuid corresponds to this patch adds resource type
and resource service columns to the slot allocations DB table.

Change-Id: I6cd7ca7f7225f8f40ba9aa8cd0eca923f5bafffa