Hardens HA router scheduler and global router creation
Sometimes a redundancy router would be scheduled to
the same hosting device as the user visible router
(or other redundancy routers for that user visible
router).

The reason for the bug is that the binding between
the redundancy router and its user visible router
is sometimes not persisted to the DB before
the scheduling of the redundancy router is attempted.

This patch solves this issue by aborting the
scheduling when this happens and keeping the redundancy
router in the backlog for later scheduling attempts.

This patch also fixes so that global routers are
created when a redundancy router with gateway set is
scheduled to a hosting device not already hosting
a neutron router with gateway set. The deletion of a
global router on a hosting devices not hosting any
neutron router with gateway set is still broken.
This will be fixed in a subsequent patch.

Fixes defects: DE785, DE793

Change-Id: I4c8203498b6ce7f6cd5a3f082c494b07aa5966c4