Fixes broken global and logical global router deletion
The deletion of a global router on a hosting device
not hosting any neutron router with gateway set was
not working. The root cause of the problem was that
the deletion of logical global routers failed. This
was because the logical global router has the global
routers as its redundancy routers. These redundancy
routers are explicitely deleted by the asr1k router
type driver. However, when the logical global router
is deleted the HA-aware router delete function also
tries to delete the redundancy routers. Alas, the
failure since the redundancy router is already
removed then.

This patch fixes that problem by using the parent
router delete function for the deletion of the
logical global router. The patch also fixes some
pep8 complaints.

Fixes defects: DE794

Change-Id: Ia37fb88e14c0506b05af6f1b905b55b54ddd96cd