Makes ha_enabled_by_default setting part of router type
So far the setting was global across all router types
through the config option 'ha_enabled_by_default' in
the 'ha' group. This is undesirable since a router type
may not support HA. This patch removes the
'ha_enabled_by_default' config option  from the ha group.
It is instead made part of the 'routertype' resource with
a default value of False. Hence, HA has to be explicitely
enabled when the router type is created (or through a
routertype update).

The patch also does some cosmetic changes and fixes
some UTs for the config agent that started failing by
changes made in an earlier patch.

Change-Id: I5926e79aca03fb4d0f3b24f50f1861f9fdc8257d