Send router update notifications when port associated with fip is deleted
The cisco l3 router plugin relied on the '_notify_routers_callback'
callback in l3_db. That callback function does not consider any HA
redundancy routers in the notification it sends. As a result, NAT
configurations in the ASRs hosting the redundancy routers were not
notified about FIP changes.

This patch solves this issue by replacing the callback handler with
one that is aware of HA redundancy routers.

Fixes defect: DE803

Change-Id: I8aba85d2a995376868801901c5ed56bdc6fba0df